TypeChecker
- Added	`isValidated(Instruction)` and `markAsvalidated(Instruction)`
- The `validate(InstrCtx, Instruction)` function now marks instructions that just finished
being validated, as, well, "validated".	It also	checks if an instruction was validtaed on
entry and if so	returns	early.
- Fixed	the `InstrCtx` and respective `Container` to provide to	it when
validating the `Value`-based argument instructions for `FuncCallInstr`
- Added	validation support for `FetchValueVar`
- We now validate later	when handling `VariableExpression`
- When handling	a `Variable` in	`<container>.<variable>` we now	will generate
a new `FetchValueVar` instruction with the `Context` from the `BinOpExpr` but
with the name being that of `<containerName>.<varname>`; we then perform
validation on this instruction
- After	validating a `FuncCallInstr` in	the case of `<containerName>.func()`
we update the name/target of the `FuncCallInstr` to that of `"<containerName>.func"`
- Fixed	bug where we called `validate(InstrCtx,	Instruction)` on a `null` instruction when
handling, presumably, the "else	case" of an `IFStmt`'s branches