Dependency
- When doing a `generalPass(Container, Context)`, if we	reach a	`Struct` entity	then
do not process.	However, since this is done, we	won't have a later chance to set
the `Context` on it, therefore do it now. This is required else	we get errors later,
specifically in	`DGen` which is	currently making use of	that `Context` it seems	(either
indirectly from	a `Instruction` copy-across during codegen or directly)