TypeChecker
- Implemented `isArrayType(Type)`, `isClassType(Type)`,	`isStructType(Type)` and `isScalarType(Type)`
- When handling	`VariableExpression`, store the	context	into `ctx` and assert it. Just cleaner code.
- When handling	the binary operators, and we get a `SymbolType.DOT`, determine if the entity referred
to by the left-hand operand is a `TypedEntity`,	then if	so, determine if the `Type` is a struct	type.
If the latter case is `true`, then update the `vLhsInstr` to a `StructDataRef` instruction
- Removed some eror checking code that appeared	after the above	change
- If the `vRhsInstr` is	a `FetchValueVar` then,	in the case whereby our	`vLhsInstr` was	a `StructDataRef`
(as explained earlier) then we will see	if the right-hand entity is a member of	that struct type, if
this IS	the case then we return	early with a `StructMemberRefInstr` that we create
- If the `vLhsIsntr` was a `StructMemberRefInstr`, meaning a reference to a member of a struct (computed
at an earlier codegen cycle) then we do	a similar thing, with nested references, we handle the case
where the `vRhsInstr` is a `FetchValueVar`, this accounts for `a.b.x` where `a`	is a struct type,
`b` is a member	of the struct referred 	to by `a` AND is a struct-typed	member variable, and then `x`
is a member variable of	the struct type	of `b`.
- Added	error handling for `StaticVariableDeclaration` dependency-node.	We only	allow assignments
for scalar types but not class types or	struct	types
- Added	support	for `StructMemberRefInstr` handling in the `Assignment_V2` case
- `beginCheck()` now checks for	cylic struct type definitions; added a unittest	for this too

TypeChecker (unittests)

- Added	two unittests for cycle	detection (one uni-module one multi-module)
- Added	a unittest for ill-member references of	struct typed entities