TypeChecker
- Implemented `isModule(Entity)`
- When doing code generation on	a `VariableExpression` AST node, we do a lookup	within
the current `Context`'s	container to see if it refers to a variable therein that
is a `Struct`-type. Note, if this fails	we do not throw	any sort of error.
- Moved	code for handling the generation of `StructMemberRefInstr`(s)
- If we	have `<modName>.<thing>` then we perform a lookup in the `<modName>`
of the `Targetable`'s target, and we do	the needed type	setting	of the instructions
here BUT we actually then update the target to `<modName>.currentTraget`